requirements.txt
accelerate
einops
ema-pytorch
numpy==1.22.4
tqdm
wandb==0.14.2
opencv-python
h5py
imageio==2.27.0
matplotlib
scikit-image
scipy
open-clip-torch
clip
lpips
omegaconf
torch-efficient-distloss
torchtyping
typeguard
hydra-core
jaxtyping
scikit-learn
timm==1.0.15
imageio-ffmpeg
transformers
sentencepiece
wheel
lightning
black
ruff
ipdb
beartype
colorama
colorspacious
moviepy
git+https://github.com/yifanyin11/latent-gaussian-rasterization.git
dacite
e3nn
plyfile
tabulate
svg.py
diffusers
rotary_embedding_torch
roma
pandas
open3d
openai
peft
qwen-vl-utils
google-genai
supervision
sentencepiece